separated ket dm from base
