+++ mrmustard/lab_dev/states/base.py
@@ -26,18 +26,16 @@ representation.
 from __future__ import annotations
 
 from abc import abstractmethod
-from itertools import product
 from typing import Sequence
 
 from enum import Enum
-import warnings
 
 import numpy as np
 from IPython.display import display
 from plotly.subplots import make_subplots
 import plotly.graph_objects as go
 
-from mrmustard import math, settings, widgets
+from mrmustard import math, settings
 from mrmustard.physics.fock import quadrature_distribution
 from mrmustard.physics.wigner import wigner_discretized
 from mrmustard.utils.typing import (
@@ -46,15 +44,7 @@ from mrmustard.utils.typing import (
     ComplexTensor,
     ComplexVector,
     RealVector,
-    Scalar,
-    Vector,
 )
-from mrmustard.physics.bargmann import (
-    wigner_to_bargmann_psi,
-    wigner_to_bargmann_rho,
-)
-from mrmustard.math.lattice.strategies.vanilla import autoshape_numba
-from mrmustard.physics.gaussian import purity
 from mrmustard.physics.representations import Bargmann, Fock
 from mrmustard.lab_dev.utils import shape_check
 from mrmustard.physics.ansatze import (
@@ -62,9 +52,8 @@ from mrmustard.physics.ansatze import (
 )
 from mrmustard.lab_dev.circuit_components_utils import BtoPS, BtoQ, TraceOut
 from mrmustard.lab_dev.circuit_components import CircuitComponent
-from mrmustard.lab_dev.wires import Wires
 
-__all__ = ["State", "DM", "Ket"]
+__all__ = ["State"]
 
 # ~~~~~~~
 # Helpers
@@ -185,7 +174,7 @@ class State(CircuitComponent):
 
             >>> from mrmustard.physics.representations import Bargmann
             >>> from mrmustard.physics.triples import coherent_state_Abc
-            >>> from mrmustard.lab_dev.states.base import Ket
+            >>> from mrmustard.lab_dev.states.ket import Ket
 
             >>> modes = [0, 1]
             >>> triple = coherent_state_Abc(x=[0.1, 0.2])  # parallel coherent states
@@ -258,7 +247,7 @@ class State(CircuitComponent):
         means: ComplexMatrix,
         name: str | None = None,
         atol_purity: float | None = 1e-5,
-    ) -> Ket | DM:  # pylint: disable=abstract-method
+    ) -> State:  # pylint: disable=abstract-method
         r"""
         Initializes a state from the covariance matrix and the vector of means of a state in
         phase space.
@@ -313,30 +302,6 @@ class State(CircuitComponent):
                 with the number of modes.
         """
 
-    def fock_distribution(self, cutoff: int) -> ComplexTensor:
-        r"""
-        Returns the Fock distribution of the state up to some cutoff.
-
-        Args:
-            cutoff: The photon cutoff.
-
-        Returns:
-            The Fock distribution.
-        """
-        fock_array = self.fock(cutoff)
-        if isinstance(self, Ket):
-            probs = (
-                math.astensor(
-                    [fock_array[ns] for ns in product(list(range(cutoff)), repeat=self.n_modes)]
-                )
-                ** 2
-            )
-        else:
-            probs = math.astensor(
-                [fock_array[ns * 2] for ns in product(list(range(cutoff)), repeat=self.n_modes)]
-            )
-        return probs
-
     def phase_space(self, s: float) -> tuple:
         r"""
         Returns the phase space parametrization of a state, consisting in a covariance matrix, a vector of means and a scaling coefficient. When a state is a linear superposition of Gaussians, each of cov, means, coeff are arranged in a batch.
@@ -356,32 +321,6 @@ class State(CircuitComponent):
         return bargmann_Abc_to_phasespace_cov_means(
             *new_state.bargmann_triple(batched=True), batched=True
         )
-
-    def quadrature_distribution(self, quad: Vector, phi: float = 0.0) -> ComplexTensor:
-        r"""
-        The (discretized) quadrature distribution of the State.
-
-        Args:
-            quad: the discretized quadrature axis over which the distribution is computed.
-            phi: The quadrature angle. ``phi=0`` corresponds to the x quadrature,
-                    ``phi=pi/2`` to the p quadrature. The default value is ``0``.
-        Returns:
-            The quadrature distribution.
-        """
-        quad = math.astensor(quad)
-        if len(quad.shape) != 1 and len(quad.shape) != self.n_modes:
-            raise ValueError(
-                "The dimensionality of quad should be 1, or match the number of modes."
-            )
-
-        if len(quad.shape) == 1:
-            quad = math.astensor(list(product(quad, repeat=len(self.modes))))
-
-        if isinstance(self, Ket):
-            return math.abs(self.quadrature(quad, phi)) ** 2
-        else:
-            quad = math.tile(quad, (1, 2))
-            return self.quadrature(quad, phi)
 
     def visualize_2d(
         self,
@@ -424,7 +363,7 @@ class State(CircuitComponent):
             raise ValueError("2D visualization not available for multi-mode states.")
         shape = [max(min_shape, d) for d in self.auto_shape()]
         state = self.to_fock(tuple(shape))
-        state = state if isinstance(state, DM) else state.dm()
+        state = state.dm()
         dm = math.sum(state.representation.array, axes=[0])
 
         x, prob_x = quadrature_distribution(dm)
@@ -540,7 +479,7 @@ class State(CircuitComponent):
             raise ValueError("3D visualization not available for multi-mode states.")
         shape = [max(min_shape, d) for d in self.auto_shape()]
         state = self.to_fock(tuple(shape))
-        state = state if isinstance(state, DM) else state.dm()
+        state = state.dm()
         dm = math.sum(state.representation.array, axes=[0])
 
         xvec = np.linspace(*xbounds, resolution)
@@ -614,7 +553,7 @@ class State(CircuitComponent):
         if self.n_modes != 1:
             raise ValueError("DM visualization not available for multi-mode states.")
         state = self.to_fock(cutoff)
-        state = state if isinstance(state, DM) else state.dm()
+        state = state.dm()
         dm = math.sum(state.representation.array, axes=[0])
 
         fig = go.Figure(
@@ -631,635 +570,3 @@ class State(CircuitComponent):
         if return_fig:
             return fig
         display(fig)
-
-    def _ipython_display_(self):  # pragma: no cover
-        is_ket = isinstance(self, Ket)
-        is_fock = isinstance(self.representation, Fock)
-        display(widgets.state(self, is_ket=is_ket, is_fock=is_fock))
-
-
-class DM(State):
-    r"""
-    Base class for density matrices.
-
-    Args:
-        modes: The modes of this density matrix.
-        representation: The representation of this density matrix.
-        name: The name of this density matrix.
-    """
-
-    short_name = "DM"
-
-    def __init__(
-        self,
-        modes: Sequence[int] = (),
-        representation: Bargmann | Fock | None = None,
-        name: str | None = None,
-    ):
-        if representation and representation.ansatz.num_vars != 2 * len(modes):
-            raise ValueError(
-                f"Expected a representation with {2*len(modes)} variables, found {representation.ansatz.num_vars}."
-            )
-        super().__init__(
-            wires=[modes, (), modes, ()],
-            name=name,
-        )
-        if representation is not None:
-            self._representation = representation
-
-    @property
-    def is_positive(self) -> bool:
-        r"""
-        Whether this DM is a positive operator.
-        """
-        batch_dim = self.representation.ansatz.batch_size
-        if batch_dim > 1:
-            raise ValueError(
-                "Physicality conditions are not implemented for batch dimension larger than 1."
-            )
-        A = self.representation.A[0]
-        m = A.shape[-1] // 2
-        gamma_A = A[:m, m:]
-
-        if (
-            math.real(math.norm(gamma_A - math.conj(gamma_A.T))) > settings.ATOL
-        ):  # checks if gamma_A is Hermitian
-            return False
-
-        return all(math.real(math.eigvals(gamma_A)) >= 0)
-
-    @property
-    def is_physical(self) -> bool:
-        r"""
-        Whether this DM is a physical density operator.
-        """
-        return self.is_positive and math.allclose(self.probability, 1, settings.ATOL)
-
-    @property
-    def probability(self) -> float:
-        r"""Probability (trace) of this DM, using the batch dimension of the Ansatz
-        as a convex combination of states."""
-        return math.sum(self._probabilities)
-
-    @property
-    def purity(self) -> float:
-        return self.L2_norm
-
-    @property
-    def _probabilities(self) -> RealVector:
-        r"""Element-wise probabilities along the batch dimension of this DM.
-        Useful for cases where the batch dimension does not mean a convex combination of states.
-        """
-        idx_ket = self.wires.output.ket.indices
-        idx_bra = self.wires.output.bra.indices
-        rep = self.representation.trace(idx_ket, idx_bra)
-        return math.real(math.sum(rep.scalar))
-
-    @property
-    def _purities(self) -> RealVector:
-        r"""Element-wise purities along the batch dimension of this DM.
-        Useful for cases where the batch dimension does not mean a convex combination of states.
-        """
-        return self._L2_norms / self._probabilities
-
-    @classmethod
-    def from_bargmann(
-        cls,
-        modes: Sequence[int],
-        triple: tuple[ComplexMatrix, ComplexVector, complex],
-        name: str | None = None,
-    ) -> State:
-        return DM(modes, Bargmann(*triple), name)
-
-    @classmethod
-    def from_fock(
-        cls,
-        modes: Sequence[int],
-        array: ComplexTensor,
-        name: str | None = None,
-        batched: bool = False,
-    ) -> State:
-        return DM(modes, Fock(array, batched), name)
-
-    @classmethod
-    def from_phase_space(
-        cls,
-        modes: Sequence[int],
-        triple: tuple,
-        name: str | None = None,
-        s: float = 0,  # pylint: disable=unused-argument
-    ) -> DM:
-        r"""
-        Initializes a density matrix from the covariance matrix, vector of means and a coefficient,
-        which parametrize the s-parametrized phase space function
-        :math:`coeff * exp(-1/2(x-means)^T cov^{-1} (x-means))`.h:`coeff * exp((x-means)^T cov^{-1} (x-means))`.
-
-
-        Args:
-            modes: The modes of this states.
-            triple: The ``(cov, means, coeff)`` triple.
-            name: The name of this state.
-            s: The phase space parameter, defaults to 0 (Wigner).
-        """
-        cov, means, coeff = triple
-        cov = math.astensor(cov)
-        means = math.astensor(means)
-        shape_check(cov, means, 2 * len(modes), "Phase space")
-        return coeff * DM(
-            modes,
-            Bargmann.from_function(fn=wigner_to_bargmann_rho, cov=cov, means=means),
-            name,
-        )
-
-    @classmethod
-    def from_quadrature(
-        cls,
-        modes: Sequence[int],
-        triple: tuple[ComplexMatrix, ComplexVector, complex],
-        phi: float = 0.0,
-        name: str | None = None,
-    ) -> State:
-        r"""
-        Initializes a state from a triple (A,b,c) that parametrizes the wavefunction
-        as `c * exp(0.5 z^T A z + b^T z)` in the quadrature representation.
-
-        Args:
-            modes: The modes of this state.
-            triple: The ``(A, b, c)`` triple.
-            phi: The angle of the quadrature. 0 corresponds to the x quadrature (default).
-            name: The name of this state.
-
-        Returns:
-            A state of type ``cls``.
-
-        Raises:
-            ValueError: If the given triple has shapes that are inconsistent
-                with the number of modes.
-        """
-        QtoB = BtoQ(modes, phi).inverse()
-        Q = DM(modes, Bargmann(*triple))
-        return DM(modes, (Q >> QtoB).representation, name)
-
-    @classmethod
-    def random(cls, modes: Sequence[int], m: int | None = None, max_r: float = 1.0) -> DM:
-        r"""
-        Samples a random density matrix. The final state has zero displacement.
-
-        Args:
-        modes: the modes where the state is defined over
-        m: is the number modes to be considered for tracing out from a random pure state (Ket)
-        if not specified, m is considered to be len(modes)
-        """
-        if m is None:
-            m = len(modes)
-
-        max_idx = max(modes)
-
-        ancilla = list(range(max_idx + 1, max_idx + m + 1))
-        full_wires = list(modes) + ancilla
-
-        psi = Ket.random(full_wires, max_r)
-        return psi[modes]
-
-    def auto_shape(
-        self, max_prob=None, max_shape=None, respect_manual_shape=True
-    ) -> tuple[int, ...]:
-        r"""
-        A good enough estimate of the Fock shape of this DM, defined as the shape of the Fock
-        array (batch excluded) if it exists, and if it doesn't exist it is computed as the shape
-        that captures at least ``settings.AUTOSHAPE_PROBABILITY`` of the probability mass of each
-        single-mode marginal (default 99.9%).
-        If the ``respect_manual_shape`` flag is set to ``True``, auto_shape will respect the
-        non-None values in ``manual_shape``.
-
-        Args:
-            max_prob: The maximum probability mass to capture in the shape (default in ``settings.AUTOSHAPE_PROBABILITY``).
-            max_shape: The maximum shape to compute (default in ``settings.AUTOSHAPE_MAX``).
-            respect_manual_shape: Whether to respect the non-None values in ``manual_shape``.
-        """
-        # experimental:
-        if self.representation.ansatz.batch_size == 1:
-            try:  # fock
-                shape = self._representation.array.shape[1:]
-            except AttributeError:  # bargmann
-                if self.representation.ansatz.polynomial_shape[0] == 0:
-                    repr = self.representation
-                    A, b, c = repr.A[0], repr.b[0], repr.c[0]
-                    repr = repr / self.probability
-                    shape = autoshape_numba(
-                        math.asnumpy(A),
-                        math.asnumpy(b),
-                        math.asnumpy(c),
-                        max_prob or settings.AUTOSHAPE_PROBABILITY,
-                        max_shape or settings.AUTOSHAPE_MAX,
-                    )
-                    shape = tuple(shape) + tuple(shape)
-                else:
-                    shape = [settings.AUTOSHAPE_MAX] * 2 * len(self.modes)
-        else:
-            warnings.warn("auto_shape only looks at the shape of the first element of the batch.")
-            shape = [settings.AUTOSHAPE_MAX] * 2 * len(self.modes)
-        if respect_manual_shape:
-            return tuple(c or s for c, s in zip(self.manual_shape, shape))
-        return tuple(shape)
-
-    def dm(self) -> DM:
-        r"""
-        The ``DM`` object obtained from this ``DM``.
-        """
-        return self
-
-    def expectation(self, operator: CircuitComponent):
-        r"""
-        The expectation value of an operator with respect to this DM.
-
-        Given the operator `O`, this function returns :math:`Tr\big(\rho O)`\, where :math:`\rho`
-        is the density matrix of this state.
-
-        The ``operator`` is expected to be a component with ket-like wires (i.e., output wires on
-        the ket side), density matrix-like wires (output wires on both ket and bra sides), or
-        unitary-like wires (input and output wires on the ket side).
-
-        Args:
-            operator: A ket-like, density-matrix like, or unitary-like circuit component.
-
-        Raise:
-            ValueError: If ``operator`` is not a ket-like, density-matrix like, or unitary-like
-                component.
-            ValueError: If ``operator`` is defined over a set of modes that is not a subset of the
-                modes of this state.
-        """
-        op_type, msg = _validate_operator(operator)
-        if op_type is OperatorType.INVALID_TYPE:
-            raise ValueError(msg)
-
-        if not operator.wires.modes.issubset(self.wires.modes):
-            msg = f"Expected an operator defined on a subset of modes `{self.modes}`, "
-            msg += f"found one defined on `{operator.modes}.`"
-            raise ValueError(msg)
-
-        leftover_modes = self.wires.modes - operator.wires.modes
-        if op_type is OperatorType.KET_LIKE:
-            result = self >> operator.dual
-            if leftover_modes:
-                result >>= TraceOut(leftover_modes)
-        elif op_type is OperatorType.DM_LIKE:
-            result = self >> operator.dual
-            if leftover_modes:
-                result >>= TraceOut(leftover_modes)
-        else:
-            result = (self @ operator) >> TraceOut(self.modes)
-
-        return result
-
-    def normalize(self) -> DM:
-        r"""
-        Returns a rescaled version of the state such that its probability is 1.
-        """
-        return self / self.probability
-
-    def __getitem__(self, modes: int | Sequence[int]) -> State:
-        r"""
-        Traces out all the modes except those given.
-        The result is returned with modes in increasing order.
-        """
-        if isinstance(modes, int):
-            modes = [modes]
-        modes = set(modes)
-
-        if not modes.issubset(self.modes):
-            msg = f"Expected a subset of `{self.modes}, found `{list(modes)}`."
-            raise ValueError(msg)
-
-        if self._parameter_set:
-            # if ``self`` has a parameter set it means it is a built-in state,
-            # in which case we slice the parameters
-            return self._getitem_builtin(modes)
-
-        # if ``self`` has no parameter set it is not a built-in state,
-        # in which case we trace the representation
-        wires = Wires(modes_out_bra=modes, modes_out_ket=modes)
-
-        idxz = [i for i, m in enumerate(self.modes) if m not in modes]
-        idxz_conj = [i + len(self.modes) for i, m in enumerate(self.modes) if m not in modes]
-        representation = self.representation.trace(idxz, idxz_conj)
-
-        return self.__class__._from_attributes(
-            representation, wires, self.name
-        )  # pylint: disable=protected-access
-
-    def __rshift__(self, other: CircuitComponent) -> CircuitComponent:
-        r"""
-        Contracts ``self`` and ``other`` (output of self into the inputs of other),
-        adding the adjoints when they are missing. Given this is a ``DM`` object which
-        has both ket and bra wires at the output, expressions like ``dm >> u`` where
-        ``u`` is a unitary will automatically apply the adjoint of ``u`` on the bra side.
-
-        Returns a ``DM`` when the wires of the resulting components are compatible with
-        those of a ``DM``, a ``CircuitComponent`` otherwise, and a scalar if there are no wires left.
-        """
-        result = super().__rshift__(other)
-        if not isinstance(result, CircuitComponent):
-            return result  # scalar case handled here
-
-        w = result.wires
-        if not w.input and w.bra.modes == w.ket.modes:
-            return DM(w.modes, result.representation)
-        return result
-
-
-class Ket(State):
-    r"""
-    Base class for all Hilbert space vectors.
-
-    Arguments:
-        modes: The modes of this ket.
-        representation: The representation of this ket.
-        name: The name of this ket.
-    """
-
-    short_name = "Ket"
-
-    def __init__(
-        self,
-        modes: Sequence[int] = (),
-        representation: Bargmann | Fock | None = None,
-        name: str | None = None,
-    ):
-        if representation and representation.ansatz.num_vars != len(modes):
-            raise ValueError(
-                f"Expected a representation with {len(modes)} variables, found {representation.ansatz.num_vars}."
-            )
-        super().__init__(
-            wires=[(), (), modes, ()],
-            name=name,
-        )
-        if representation is not None:
-            self._representation = representation
-
-    @property
-    def is_physical(self) -> bool:
-        r"""
-        Whether the ket object is a physical one.
-        """
-        batch_dim = self.representation.ansatz.batch_size
-        if batch_dim > 1:
-            raise ValueError(
-                "Physicality conditions are not implemented for batch dimension larger than 1."
-            )
-
-        A = self.representation.A[0]
-
-        return all(math.abs(math.eigvals(A)) < 1) and math.allclose(
-            self.probability, 1, settings.ATOL
-        )
-
-    @property
-    def probability(self) -> float:
-        r"""Probability of this Ket (L2 norm squared)."""
-        return self.L2_norm
-
-    @property
-    def purity(self) -> float:
-        return 1.0
-
-    @property
-    def _probabilities(self) -> RealVector:
-        r"""Element-wise L2 norm squared along the batch dimension of this Ket."""
-        return self._L2_norms
-
-    @classmethod
-    def from_bargmann(
-        cls,
-        modes: Sequence[int],
-        triple: tuple[ComplexMatrix, ComplexVector, complex],
-        name: str | None = None,
-    ) -> State:
-        return Ket(modes, Bargmann(*triple), name)
-
-    @classmethod
-    def from_fock(
-        cls,
-        modes: Sequence[int],
-        array: ComplexTensor,
-        name: str | None = None,
-        batched: bool = False,
-    ) -> State:
-        return Ket(modes, Fock(array, batched), name)
-
-    @classmethod
-    def from_phase_space(
-        cls,
-        modes: Sequence[int],
-        triple: tuple,
-        name: str | None = None,
-        atol_purity: float | None = 1e-5,
-    ) -> Ket:
-        cov, means, coeff = triple
-        cov = math.astensor(cov)
-        means = math.astensor(means)
-        shape_check(cov, means, 2 * len(modes), "Phase space")
-        if atol_purity:
-            p = purity(cov)
-            if p < 1.0 - atol_purity:
-                msg = f"Cannot initialize a Ket: purity is {p:.5f} (must be at least 1.0-{atol_purity})."
-                raise ValueError(msg)
-        return Ket(
-            modes,
-            coeff * Bargmann.from_function(fn=wigner_to_bargmann_psi, cov=cov, means=means),
-            name,
-        )
-
-    @classmethod
-    def from_quadrature(
-        cls,
-        modes: Sequence[int],
-        triple: tuple[ComplexMatrix, ComplexVector, complex],
-        phi: float = 0.0,
-        name: str | None = None,
-    ) -> State:
-        QtoB = BtoQ(modes, phi).inverse()
-        Q = Ket(modes, Bargmann(*triple))
-        return Ket(modes, (Q >> QtoB).representation, name)
-
-    @classmethod
-    def random(cls, modes: Sequence[int], max_r: float = 1.0) -> Ket:
-        r"""
-        Generates a random zero displacement state.
-
-        Args:
-            modes: The modes of the state.
-            max_r: Maximum squeezing parameter over which we make random choices.
-        Output is a Ket
-        """
-
-        m = len(modes)
-        S = math.random_symplectic(m, max_r)
-        transformation = (
-            1
-            / np.sqrt(2)
-            * math.block(
-                [
-                    [
-                        math.eye(m, dtype=math.complex128),
-                        math.eye(m, dtype=math.complex128),
-                    ],
-                    [
-                        -1j * math.eye(m, dtype=math.complex128),
-                        1j * math.eye(m, dtype=math.complex128),
-                    ],
-                ]
-            )
-        )
-        S = math.conj(math.transpose(transformation)) @ S @ transformation
-        S_1 = S[:m, :m]
-        S_2 = S[:m, m:]
-        A = S_2 @ math.conj(math.inv(S_1))  # use solve for inverse
-        b = math.zeros(m, dtype=A.dtype)
-        psi = cls.from_bargmann(modes, [[A], [b], [complex(1)]])
-        return psi.normalize()
-
-    def auto_shape(
-        self, max_prob=None, max_shape=None, respect_manual_shape=True
-    ) -> tuple[int, ...]:
-        r"""
-        A good enough estimate of the Fock shape of this Ket, defined as the shape of the Fock
-        array (batch excluded) if it exists, and if it doesn't exist it is computed as the shape
-        that captures at least ``settings.AUTOSHAPE_PROBABILITY`` of the probability mass of each
-        single-mode marginal (default 99.9%).
-        If the ``respect_manual_shape`` flag is set to ``True``, auto_shape will respect the
-        non-None values in ``manual_shape``.
-
-        Args:
-            max_prob: The maximum probability mass to capture in the shape (default from ``settings.AUTOSHAPE_PROBABILITY``).
-            max_shape: The maximum shape to compute (default from ``settings.AUTOSHAPE_MAX``).
-            respect_manual_shape: Whether to respect the non-None values in ``manual_shape``.
-        """
-        # experimental:
-        if self.representation.ansatz.batch_size == 1:
-            try:  # fock
-                shape = self._representation.array.shape[1:]
-            except AttributeError:  # bargmann
-                if self.representation.ansatz.polynomial_shape[0] == 0:
-                    repr = self.representation.conj() & self.representation
-                    A, b, c = repr.A[0], repr.b[0], repr.c[0]
-                    repr = repr / self.probability
-                    shape = autoshape_numba(
-                        math.asnumpy(A),
-                        math.asnumpy(b),
-                        math.asnumpy(c),
-                        max_prob or settings.AUTOSHAPE_PROBABILITY,
-                        max_shape or settings.AUTOSHAPE_MAX,
-                    )
-                else:
-                    shape = [settings.AUTOSHAPE_MAX] * len(self.modes)
-        else:
-            warnings.warn("auto_shape only looks at the shape of the first element of the batch.")
-            shape = [settings.AUTOSHAPE_MAX] * len(self.modes)
-        if respect_manual_shape:
-            return tuple(c or s for c, s in zip(self.manual_shape, shape))
-        return tuple(shape)
-
-    def dm(self) -> DM:
-        r"""
-        The ``DM`` object obtained from this ``Ket``.
-        """
-        dm = self @ self.adjoint
-        ret = DM._from_attributes(dm.representation, dm.wires, self.name)
-        ret.manual_shape = self.manual_shape + self.manual_shape
-        return ret
-
-    def expectation(self, operator: CircuitComponent):
-        r"""
-        The expectation value of an operator calculated with respect to this Ket.
-
-        Given the operator `O`, this function returns :math:`Tr\big(|\psi\rangle\langle\psi| O)`\,
-        where :math:`|\psi\rangle` is the vector representing this state.
-
-        The ``operator`` is expected to be a component with ket-like wires (i.e., output wires on
-        the ket side), density matrix-like wires (output wires on both ket and bra sides), or
-        unitary-like wires (input and output wires on the ket side).
-
-        Args:
-            operator: A ket-like, density-matrix like, or unitary-like circuit component.
-
-        Raise:
-            ValueError: If ``operator`` is not a ket-like, density-matrix like, or unitary-like
-                component.
-            ValueError: If ``operator`` is defined over a set of modes that is not a subset of the
-                modes of this state.
-        """
-        op_type, msg = _validate_operator(operator)
-        if op_type is OperatorType.INVALID_TYPE:
-            raise ValueError(msg)
-
-        if not operator.wires.modes.issubset(self.wires.modes):
-            msg = f"Expected an operator defined on a subset of modes `{self.modes}`, "
-            msg += f"found one defined on `{operator.modes}.`"
-            raise ValueError(msg)
-
-        leftover_modes = self.wires.modes - operator.wires.modes
-        if op_type is OperatorType.KET_LIKE:
-            result = self @ operator.dual
-            result @= result.adjoint
-            result >>= TraceOut(leftover_modes)
-
-        elif op_type is OperatorType.DM_LIKE:
-            result = (self.adjoint @ (self @ operator.dual)) >> TraceOut(leftover_modes)
-
-        else:
-            result = (self @ operator) >> self.dual
-
-        return result
-
-    def normalize(self) -> Ket:
-        r"""
-        Returns a rescaled version of the state such that its probability is 1
-        """
-        return self / math.sqrt(self.probability)
-
-    def __getitem__(self, modes: int | Sequence[int]) -> State:
-        r"""
-        Reduced density matrix obtained by tracing out all the modes except those in the given
-        ``modes``. Note that the result is returned with modes in increasing order.
-        """
-        if isinstance(modes, int):
-            modes = [modes]
-        modes = set(modes)
-
-        if not modes.issubset(self.modes):
-            raise ValueError(f"Expected a subset of `{self.modes}, found `{list(modes)}`.")
-
-        if self._parameter_set:
-            # if ``self`` has a parameter set, it is a built-in state, and we slice the
-            # parameters
-            return self._getitem_builtin(modes)
-
-        # if ``self`` has no parameter set, it is not a built-in state.
-        # we must turn it into a density matrix and slice the representation
-        return self.dm()[modes]
-
-    def __rshift__(self, other: CircuitComponent | Scalar) -> CircuitComponent | Batch[Scalar]:
-        r"""
-        Contracts ``self`` and ``other`` (output of self into the inputs of other),
-        adding the adjoints when they are missing. Given this is a ``Ket`` object which
-        has only ket wires at the output, in expressions like ``ket >> channel`` where ``channel``
-        has wires on the ket and bra sides the adjoint of ket is automatically added, effectively
-        calling ``ket.adjoint @ (ket @ channel)`` and the method returns a new ``DM``.
-        In expressions lke ``ket >> u`` where ``u`` is a unitary, the adjoint of ``ket`` is
-        not needed and the method returns a new ``Ket``.
-
-        Returns a ``DM`` or a ``Ket`` when the wires of the resulting components are compatible
-        with those of a ``DM`` or of a ``Ket``. Returns a ``CircuitComponent`` in general,
-        and a (batched) scalar if there are no wires left, for convenience.
-        """
-        result = super().__rshift__(other)
-        if not isinstance(result, CircuitComponent):
-            return result  # scalar case handled here
-
-        if not result.wires.input:
-            if not result.wires.bra:
-                return Ket(result.wires.modes, result.representation)
-            elif result.wires.bra.modes == result.wires.ket.modes:
-                result = DM(result.wires.modes, result.representation)
-        return result
