+++ mrmustard/lab_dev/states/squeezed_vacuum.py
@@ -22,7 +22,7 @@ from typing import Sequence
 
 from mrmustard.physics.representations import Bargmann
 from mrmustard.physics import triples
-from .base import Ket
+from .ket import Ket
 from ..utils import make_parameter, reshape_params
 
 __all__ = ["SqueezedVacuum"]
