+++ mrmustard/lab_dev/states/vacuum.py
@@ -22,7 +22,7 @@ from typing import Sequence
 
 from mrmustard.physics.representations import Bargmann
 from mrmustard.physics import triples
-from .base import Ket
+from .ket import Ket
 
 __all__ = ["Vacuum"]
